Added the Sumi and Refactored all code
