+++ gradle.properties
@@ -25,11 +25,11 @@ geckolib_version=4.4.7
 sbl_version=1.14.2
 
 # Mod Information
-mod_id=talesofvarmithore
+mod_id=tov
 mod_name=Tales of Varmithore
 mod_license=All Rights Reserved
 mod_version=1.0.0
-mod_group_id=com.nineteen.talesofvarmithore
+mod_group_id=com.tov.tov
 mod_authors=
 mod_description=
