--- src/main/java/com/nineteen/talesofvarmithore/setup/datagen/DataGenerators.java
@@ -1,20 +1,0 @@
-package com.nineteen.talesofvarmithore.setup.datagen;
-
-import com.nineteen.talesofvarmithore.TVModInfo;
-import net.minecraft.data.DataGenerator;
-import net.minecraft.data.PackOutput;
-import net.minecraftforge.common.data.ExistingFileHelper;
-import net.minecraftforge.data.event.GatherDataEvent;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.common.Mod;
-
-@Mod.EventBusSubscriber(modid = TVModInfo.MOD_ID, bus = Mod.EventBusSubscriber.Bus.MOD)
-public class DataGenerators {
-    @SubscribeEvent
-    public static void gatherData(GatherDataEvent event) {
-        DataGenerator generator = event.getGenerator();
-        PackOutput packOutput = generator.getPackOutput();
-        ExistingFileHelper existingFileHelper = event.getExistingFileHelper();
-        generator.addProvider(event.includeClient(), new ItemModelProvider(packOutput, existingFileHelper));
-    }
-}
