+++ src/main/java/com/tov/tov/common/entity/sumi/Sumi.java
@@ -1,0 +1,52 @@
+package com.tov.tov.common.entity.sumi;
+
+import com.tov.tov.api.common.TamableMob;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.Mob;
+import net.minecraft.world.entity.PathfinderMob;
+import net.minecraft.world.entity.ai.attributes.AttributeSupplier;
+import net.minecraft.world.entity.ai.attributes.Attributes;
+import net.minecraft.world.level.Level;
+import software.bernie.geckolib.animatable.GeoEntity;
+import software.bernie.geckolib.core.animatable.instance.AnimatableInstanceCache;
+import software.bernie.geckolib.core.animation.AnimatableManager;
+import software.bernie.geckolib.core.animatable.instance.SingletonAnimatableInstanceCache;
+import software.bernie.geckolib.core.animation.AnimationController;
+import software.bernie.geckolib.core.animation.AnimationState;
+import software.bernie.geckolib.core.animation.RawAnimation;
+import software.bernie.geckolib.core.object.PlayState;
+
+public class Sumi extends TamableMob implements GeoEntity  {
+    public Sumi(EntityType<? extends PathfinderMob> pEntityType, Level pLevel) {
+        super(pEntityType, pLevel);
+    }
+
+    public static AttributeSupplier createAttributes() {
+        return Mob.createMobAttributes()
+                .add(Attributes.MAX_HEALTH, 35)
+                .add(Attributes.ATTACK_DAMAGE, 9)
+                .add(Attributes.MOVEMENT_SPEED, 0.1f)
+                .build();
+    }
+
+    @Override
+    public void registerControllers(AnimatableManager.ControllerRegistrar controllers) {
+        controllers.add(new AnimationController<>(this, "main", 5, this::predicate));
+    }
+
+    private PlayState predicate(AnimationState<Sumi> event) {
+        return event.setAndContinue(RawAnimation.begin().thenLoop("sumi.swim"));
+    }
+
+    public int getVariant(){
+        return 0;
+    }
+
+
+    private final AnimatableInstanceCache cache = new SingletonAnimatableInstanceCache(this);
+
+    @Override
+    public AnimatableInstanceCache getAnimatableInstanceCache() {
+        return cache;
+    }
+}
