+++ src/main/java/com/tov/tov/setup/common/registries/RegistryLoader.java
@@ -1,0 +1,13 @@
+package com.tov.tov.setup.common.registries;
+
+import net.minecraftforge.eventbus.api.IEventBus;
+
+public final class RegistryLoader {
+
+    public static void load(IEventBus forgeBus, IEventBus modBus) {
+        TVEntities.ENTITY_TYPES.register(modBus);
+        TVItems.ITEMS.register(modBus);
+    }
+
+    private RegistryLoader(){}
+}
