--- src/main/resources/assets/talesofvarmithore/lang/en_us.json
@@ -1,4 +1,0 @@
-{
-  "item.talesofvarmithore.wilkor_spawn_egg" : "Wilkor Spawn Egg",
-  "entity.talesofvarmithore.wilkor" : "Wilkor"
-}
