+++ src/main/resources/assets/tov/animations/wilkor.animation.json
