+++ src/main/resources/assets/tov/lang/en_us.json
@@ -1,0 +1,4 @@
+{
+  "item.tov.wilkor_spawn_egg" : "Wilkor Spawn Egg",
+  "entity.tov.wilkor" : "Wilkor"
+}
