+++ src/main/resources/assets/tov/textures/entity/wilkor/wilkor_sand.png
